Possible fix for issue #53
Merged fold.py into filter.py, at FichteFoll's suggestion, to prevent ST 3 from greying out our menu items.
