+++ messages.json
@@ -7,5 +7,6 @@
 	"3.0.0": "messages/3.0.0.md",
 	"3.1.0": "messages/3.1.0.md",
 	"3.1.2": "messages/3.1.2.md",
-	"3.1.3": "messages/3.1.3.md"
+	"3.1.3": "messages/3.1.3.md",
+	"3.1.4": "messages/3.1.4.md"
 }
