+++ messages/3.1.4.md
@@ -1,0 +1,7 @@
+#3.1.4  August 8, 2019
+
+This release is for Sublime Text 3 users. I have separate versions of the plugin for Sublime Text 2 and 3, and the version for Sublime Text 2 remains unchanged for now.
+
+##Changes
+
+- Merged fold.py into filter.py as suggested by FichteFoll to resolve issue #53, which sometimes caused Filter Lines menu items to appear disabled in ST 3.
